Add a test for the stitcher
This commit sets up a set of fake images to stitch and tests stitching
them together in a perfectly side-to-side but nonoverlapping grid.

I also added some other fixes because I broke the tests in the last
stitcher commit by switching the parameters class to BaseSettings
instead of BaseModel (I switched it back.)

Tested by:
- `./dev/run_tests.sh`
- `./dev/autofix_lint.sh`
- `./dev/format.sh`
- `./dev/type_check.sh`
