+++ dev/run_tests.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -ex
+
+uv run python -m unittest \
+    image_stitcher.parameters_test \
+    image_stitcher.stitcher_test
